requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
transformers>=4.35.0
torch>=2.1.0
google-generativeai>=0.3.0
langdetect>=1.0.9
rapidfuzz>=3.5.0
python-dotenv>=1.0.0

